requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
pydantic-settings
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
python-multipart
jinja2
typer
pytest 
httpx 
pytest-asyncio
-e .